Scanner app changed
1. Added 100 ms scan intervals fr 2 seconds
2. Millisecond precision global clock introduced
3. Time synchronization needs to be implemented.

+++ ble5_simple_peripheral_cc2640r2lp_app/Application/simple_peripheral.c
@@ -419,7 +419,7 @@ static PIN_State ledPinState;
 
 //State machine maintaining state of the GAP
 static gaprole_States_t curr_state = GAPROLE_INIT;
-static unsigned int global_clock;
+static uint64_t global_clock;
 
 /*
  * Initial LED pin configuration table
